added boxed type instead of primitive type. changed to double to allow halves of hours

+++ teachedapp/src/main/java/com/teachedapp/dao/Lesson.java
@@ -28,7 +28,7 @@ public class Lesson {
     private String notes;
 
     @Column
-    private short duration; // in hours
+    private Double duration; // in hours
 
     @Column(name = "start_date")
     private Date startDate;
